:truck: Refs #263 - Move redirect view
