+++ src/openforms/appointments/api/urls.py
@@ -6,7 +6,6 @@ from .views import (
     LocationsListView,
     ProductsListView,
     TimesListView,
-    VerifyCancelAppointmentLinkView,
 )
 
 urlpatterns = [
@@ -14,11 +13,6 @@ urlpatterns = [
     path("locations", LocationsListView.as_view(), name="appointments-locations-list"),
     path("dates", DatesListView.as_view(), name="appointments-dates-list"),
     path("times", TimesListView.as_view(), name="appointments-times-list"),
-    path(
-        "<str:base64_submission_uuid>/<str:token>/verify",
-        VerifyCancelAppointmentLinkView.as_view(),
-        name="appointments-verify-cancel-appointment-link",
-    ),
     path(
         "<str:submission_uuid>/cancel",
         CancelAppointmentView.as_view(),
